requirements.txt
discord.py
pybit~=5.7.0
TA-Lib
requests~=2.32.3
pandas~=2.2.2
aiohttp
setuptools~=68.2.0